RUNOS v0.4
* Statistic service (web ui) - gathering stats from devices and send to the web interface. From web you can see blinking links: green, yellow, red depending on actual load.
* Flow manager (web ui, rest api) - caching flows in the controller. From web you can see switches' settings with flow table entries.
* STP service for broadcast packets - calculating spaning tree for a network and pushing rules for broadcast packets. Note it doesn't work for disconnected network.
* Pure cbench mode - just test packet_in/packet_out I/O. It's different from Cbench app that uses the entire TraceTree pipeline.
* Manual how to run Runos with ha-linux setup to have a cold backup instance.

+++ src/Loader.cc
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
+#include "Loader.hh"
+
 #include <vector>
 #include <unordered_map>
 #include <mutex>
 
 #include "Common.hh"
-#include "Loader.hh"
 #include "Application.hh"
 #include "Config.hh"
 
@@ -212,7 +213,7 @@ AppInfo* LoaderImpl::initialize(std::string serviceId)
     int specific_thread = config_get(config_cd(config, app->provides()),
             "pin-to-thread", -1);
 
-    if (specific_thread >= 0 && specific_thread < thread.size()) {
+    if (specific_thread >= 0 && (size_t) specific_thread < thread.size()) {
         app_thread = thread[specific_thread];
 
         VLOG(2) << "  moveToThread(" << app->provides()
